admin/requirements.txt
PyYAML==3.11
ecdsa==0.13
paramiko==1.16.0
pycrypto==2.6.1
